+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [0.5.0](https://github.com/dldevinc/paper-uploads/tree/v0.5.0) - 2021-04-07
+- Added `remove_variations` management command.
 - The `recreate_variations` management command has been completely rewritten.
 - Removed default permissions from all models.
 - Add Python 3.9 support